Mega C++ commit, sdl updates
Notes: FS crashes when loading a map during the renderer finding shader
files.
Some features are off/missing.

+++ code/botlib/be_aas_reach.cpp
@@ -591,7 +591,7 @@ int AAS_FallDamageDistance(void)
 {
 	float maxzvelocity, gravity, t;
 
-	maxzvelocity = sqrt(30 * 10000);
+	maxzvelocity = sqrtf(30 * 10000);
 	gravity = aassettings.phys_gravity;
 	t = maxzvelocity / gravity;
 	return 0.5 * gravity * t * t;
