+++ code/renderergl2/tr_extramath.h
@@ -57,7 +57,7 @@ void Matrix16SimpleInverse( const matrix_t in, matrix_t out);
 #define FloatToByte(a)          (byte)((a) * 255.0f)
 
 #define RGBtosRGB(a)            (((a) < 0.0031308f) ? (12.92f * (a)) : (1.055f * pow((a), 0.41666f) - 0.055f))
-#define sRGBtoRGB(a)            (((a) <= 0.04045f)  ? ((a) / 12.92f) : (pow((((a) + 0.055f) / 1.055f), 2.4)) )
+#define sRGBtoRGB(a)            (((a) <= 0.04045f)  ? ((a) / 12.92f) : (pow((double)(((a) + 0.055f) / 1.055f), (double)2.4)) )
 
 static ID_INLINE int VectorCompare4(const vec4_t v1, const vec4_t v2)
 {
